chore: removed unecessary integration test and added tokenRegistry test
